adding changes to import

+++ test/lib/streams.js
@@ -9,16 +9,19 @@ const { assert } = require('chai'),
       { ExportStream, ImportStream } = require('../../src/lib/stream'),
       { ExportFileAdapter } = require('../../src/lib/stream/adapters/file_adapter')
 
-describe('Export Adapter', () => {
+describe('Export and Import Adapters', () => {
 
-  let blob = null
+  let blob,
+      streamedBlob = null
 
   beforeEach(() => {
     blob = fs.createReadStream(`${process.cwd()}/test/data/blob.ndjson`)
+    streamedBlob = fs.createReadStream(`${process.cwd()}/test/data/blob_with_streams.ndjson`)
   })
 
   afterEach(() => {
     blob = null
+    streamedBlob = null
   })
 
   it('export using file adapter with default layout', (done) => {
@@ -28,11 +31,11 @@ describe('Export Adapter', () => {
           streamWriter = new ExportStream(stream, { format }),
           onEnd = (error) => {
             if (error) {
-              // rimraf.sync(tempDir)
+              rimraf.sync(tempDir)
               done(error)
             }
             glob('**/*.{yaml,js,png,jpeg,ico,gif,html,txt}', { cwd: tempDir }, (err, files) => {
-              // rimraf.sync(tempDir)
+              rimraf.sync(tempDir)
               if (err) {
                 done(err)
               } else {
@@ -45,20 +48,29 @@ describe('Export Adapter', () => {
     pump(blob, stream, streamWriter, new ExportFileAdapter(tempDir, { format }), onEnd)
   })
 
-})
+  it('export using streamIds for assets', (done) => {
+    const tempDir = path.join(process.cwd(), `output-${new Date().getTime()}`),
+          stream = ndjson.parse(),
+          format = 'yaml',
+          streamWriter = new ExportStream(stream, { format }),
+          onEnd = (error) => {
+            if (error) {
+              rimraf.sync(tempDir)
+              done(error)
+            }
+            glob('**/*.{yaml,js,png,jpeg,ico,gif,html,txt}', { cwd: tempDir }, (err, files) => {
+              rimraf.sync(tempDir)
+              if (err) {
+                done(err)
+              } else {
+                assert(files.length === 120, 'there are some missing files created')
+                done()
+              }
+            })
+          }
 
-describe('Import Adapters', () => {
-
-  let blob = null
-
-  beforeEach(() => {
-    blob = fs.createReadStream(`${process.cwd()}/test/data/blob.ndjson`)
+    pump(streamedBlob, stream, streamWriter, new ExportFileAdapter(tempDir, { format }), onEnd)
   })
-
-  afterEach(() => {
-    blob = null
-  })
-
 
   it('testing import adapter', (done) => {
 
@@ -66,6 +78,7 @@ describe('Import Adapters', () => {
           stream = ndjson.parse(),
           format = 'yaml',
           streamWriter = new ExportStream(stream, { format }),
+          fileWriter = new ExportFileAdapter(tempDir, { format }),
           onEnd = async(error) => {
             if (error) {
               rimraf.sync(tempDir)
@@ -73,29 +86,32 @@ describe('Import Adapters', () => {
             }
 
             // Update a file in order to get a blob to send
-            const exportedFile = fs.createReadStream(`${process.cwd()}/test/data/medable.jpg`)
-            exportedFile.pipe(fs.createWriteStream(`${tempDir}/env/assets/env.logo.content.jpeg`))
+            if (fs.existsSync(`${tempDir}/env/assets/env.logo.content.jpeg`)) {
+              const exportedFile = fs.createReadStream(`${process.cwd()}/test/data/medable.jpg`)
+              exportedFile.pipe(fs.createWriteStream(`${tempDir}/env/assets/env.logo.content.jpeg`))
 
-            const importAdapter = new ImportStream(tempDir, format),
-                  ndStream = ndjson.stringify(),
-                  items = []
+              const importAdapter = new ImportStream(tempDir, format),
+                    ndStream = ndjson.stringify(),
+                    items = []
 
-            ndStream.on('data', line => items.push(line))
-            pump(importAdapter, ndStream, () => {
-              rimraf.sync(tempDir)
-              const loadedItems = _.map(items, i => JSON.parse(i)),
-                    blobItems = _.filter(loadedItems, i => i.data && i.streamId),
-                    otherItems = _.filter(loadedItems, i => !i.data && !i.streamId)
-              assert(otherItems.length === 43, 'there are more/less files than loaded')
-              assert(blobItems.length === 1, 'there are more/less blob items than loaded')
-              fs.writeFileSync('/Users/gastonrobledo/test.ndjson', items.join('\n'))
-              done()
-            })
-            // call this in order to trigger end event or declare a data listener
-            ndStream.resume()
+              ndStream.on('data', line => items.push(line))
+              pump(importAdapter, ndStream, () => {
+                rimraf.sync(tempDir)
+                const loadedItems = _.map(items, i => JSON.parse(i)),
+                      blobItems = _.filter(loadedItems, i => i.data && i.streamId),
+                      otherItems = _.filter(loadedItems, i => !i.data && !i.streamId)
+                assert(otherItems.length === 43, 'there are more/less files than loaded')
+                assert(blobItems.length === 1, 'there are more/less blob items than loaded')
+                fs.writeFileSync('/Users/gastonrobledo/test.ndjson', items.join('\n'))
+                done()
+              })
+              // call this in order to trigger end event or declare a data listener
+              // ndStream.resume()
+            } else {
+              done('Error')
+            }
           }
-
-    pump(blob, stream, streamWriter, new ExportFileAdapter(tempDir, { format }), onEnd)
+    pump(blob, stream, streamWriter, fileWriter, onEnd)
   })
 
 })
